redactor: renamed and replaced some directories and updated profile page loading text and removed useFetch costume hook
