+++ frontend/src/app/gallery/gallery_component/GalleryImage.tsx
