+++ frontend/src/general_component/Logo.tsx
