+++ frontend/src/general_component/design/FloatingShapes.tsx
